LayoutTests/accessibility/loading-iframe-sends-notification.html

<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>

  <script>
    if (window.layoutTestController)
        layoutTestController.waitUntilDone();

    function runTest()
    {
        description("This tests that when an iframe finishes loading, it sends a layout complete notification.");

        if (window.accessibilityController) {
            // Initially, the iframe's webarea is empty.
            window.root = accessibilityController.rootElement;
            window.body = root.childAtIndex(0);
            window.iframe = body.childAtIndex(1).childAtIndex(0);
            window.scrollarea = iframe.childAtIndex(0);
            window.subwebarea = scrollarea.childAtIndex(0);
            shouldBeFalse("subwebarea.childrenCount > 0");

            iframe.addNotificationListener(function (notification) {
                // Make sure that we get a LayoutComplete notification and that
                // immediately after the notification is received, the iframe's
                // webarea has content.
                debug('Got notification on iframe: ' + notification);
                window.newScrollarea = iframe.childAtIndex(0);
                window.newSubwebarea = newScrollarea.childAtIndex(0);
                shouldBeTrue("newSubwebarea.childrenCount > 0");
            });
        }

        window.iframeElement = document.getElementById("iframe");
        iframeElement.addEventListener("load", function() {
            window.setTimeout(function() {
                debug('<br /><span class="pass">TEST COMPLETE</span>');
                if (window.layoutTestController)
                    layoutTestController.notifyDone();
            }, 10);
        }, false);

        // Load content into the iframe. This will trigger the event
        // handler above, which will check that the accessibility tree
        // was updated with new content.
        window.iframeElement.src = "data:text/html,<body><button>Click me</button></body>";

    }

    window.addEventListener('load', function() {
        setTimeout(runTest, 10);
    }, false);

  </script>
</head>
<body>

<p>Before</p>

<iframe id="iframe"></iframe>

<p>After</p>

<p>End of test</p>

<p id="description"></p>
<div id="console"></div>

</body>
</html>
